pyproject.toml
[project]
name = "adk-rag"
version = "0.1.0"
description = "Basic RAG agent using Google ADK and Vertex AI Search"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",
    "google-adk>=0.1.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-discoveryengine>=0.14.0",
    "pdfplumber>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
